+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `MoveFilesToDirectory` method to `HtmlValidationResult`.
 
+### Changed
+
+- Make `HtmlValidationOptions` `bool KeepHtmlFileWhenValid` property obsolete
+  and replace it with new `HtmlSaveCondition SaveHtmlToFile` property.
+
 ## [1.0.0] - 2021-06-28
 
 Initial version release.